migrate details screens
