--- frontend/src/components/details-tabs/DetailsTabs.module.scss
@@ -1,59 +1,0 @@
-@import 'styles/spacings';
-@import 'styles/colors';
-
-.cardsContainer {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  margin-bottom: $spacing-l * 2;
-
-  .smallCard {
-    display: flex;
-    position: relative;
-    justify-content: center;
-    width: 100%;
-    padding: $spacing-l;
-    align-items: center;
-    cursor: pointer;
-
-    &:hover {
-      opacity: 0.7;
-    }
-
-    &__disabled {
-      cursor: default !important;
-
-      &:hover {
-        opacity: 1;
-      }
-    }
-
-    * + * {
-      margin-top: $spacing-l;
-    }
-
-    .labelValue {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      text-align: center;
-
-      .clickable {
-        color: $blue !important;
-      }
-    }
-
-    .caret {
-      justify-self: end;
-      position: absolute;
-      right: $spacing-l;
-    }
-  }
-
-  .smallCard + .smallCard {
-    margin-left: $spacing-l;
-  }
-}
-
-.contentDetailsContainer {
-}
