+++ frontend/src/pages/events/EventsTable.module.scss
